requirements.txt
alabaster==0.7.9
appdirs==1.4.3
appnope==0.1.0
appscript==1.0.1
argcomplete==1.0.0
astroid==1.4.9
astropy==1.3
Babel==2.3.4
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.5.3
bitarray==0.8.1
bokeh==0.12.5
click==6.7
datashape==0.5.2
decorator==4.0.11
dj-database-url==0.4.2
Django==1.10.5
django-classy-tags==0.8.0
django-crispy-forms==1.6.1
django-filter==1.0.2
django-hashid-field==1.2.3
django-markdown-deux==1.0.5
django-pagedown==0.1.1
django-taggit==0.22.0
django-taggit-templatetags2==1.6.1
djangorestframework==3.6.2
Flask==0.12
Flask-Cors==3.0.2
future==0.16.0
gunicorn==19.7.1
hashids==1.2.0
itsdangerous==0.24
Jinja2==2.9.4
lazy-object-proxy==1.2.2
Markdown==2.6.8
markdown2==2.3.3
MarkupSafe==0.23
multipledispatch==0.4.9
networkx==1.11
numpy==1.11.3
odo==0.5.0
olefile==0.44
packaging==16.8
pandas==0.19.2
Pillow==4.0.0
psutil==5.0.1
psycopg2==2.7.1
pyparsing==2.2.0
python-dateutil==2.6.0
python-http-client==2.2.1
pytz==2016.10
PyYAML==3.12
requests==2.13.0
sendgrid==4.0.0
six==1.10.0
toolz==0.8.2
tornado==4.5.1
Unidecode==0.4.20
Werkzeug==0.11.15
whitenoise==3.3.0
wrapt==1.10.8
